requirements.txt
flask
pymongo
openai
python-dotenv
flask-cors
werkzeug